% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package setup and configuration}
\usage{
.onLoad(libname, pkgname)
}
\description{
These directives ensure the NAMESPACE is generated correctly
for use with Rcpp, rstan, and rstantools.
}
