% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.bedcast.R
\name{extract}
\alias{extract}
\title{Extract information from an object}
\usage{
extract(x, ...)
}
\arguments{
\item{x}{An object to extract from.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The result of extraction, depending on the class of \code{x}.
}
\description{
A generic function to extract information from objects.
}
