% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bedcast.R
\name{plot.bedcast}
\alias{plot.bedcast}
\title{Plotting generic for class bedcast}
\usage{
\method{plot}{bedcast}(x, type = c("fit", "growthrate", "parameters"), ...)
}
\arguments{
\item{x}{Object of class \code{bedcast} returned by
fit_bedcaster.}

\item{type}{What type of visualisation. "fit" will call
\code{vis_bedcast_fit}, "growthrate" will call
\code{vis_bedcast_growthrate} and "vis_bedcast_parameters" will call
\code{vis_parameters}.}

\item{...}{Other arguments passed to plotting functions.}
}
\description{
Plotting generic for class bedcast
}
