% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bedcast.R
\name{print.bedcast}
\alias{print.bedcast}
\title{Print method for bedcaster objects}
\usage{
\method{print}{bedcast}(x, ...)
}
\arguments{
\item{x}{A bedcaster object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for bedcaster objects
}
